requirements.txt
# Core requirements
Django>=5.0,<6.0
psycopg2-binary>=2.9.9
django-crispy-forms>=2.0
django-filter>=23.5
django-select2>=8.1.2
django-tables2>=2.6.0
django-bootstrap5>=23.3
Pillow>=10.0.0

# REST API
djangorestframework>=3.14.0

# Development tools
python-dotenv>=1.0.0
